Integrate lightbox with Image component
