+++ frontend/src/App.scss
@@ -13,7 +13,6 @@
 @import "./components/editCard/styles";
 @import "./components/form/styles";
 @import "./components/fragments/styles";
-@import "./components/imageGallery/styles";
 @import "./components/imageChangeRow/styles";
 @import "./components/list/styles";
 @import "./components/performerCard/styles";
