+++ frontend/src/components/image/styles.scss
@@ -37,4 +37,159 @@
       max-height: 100%;
     }
   }
+
+  &-magnify {
+    align-items: center;
+    background-color: rgba(0 0 0 / 15%);
+    color: rgba(255 255 255 / 85%);
+    display: flex;
+    inset: 0;
+    justify-content: center;
+    opacity: 0;
+    position: absolute;
+    transition: opacity 0.25s ease;
+    z-index: 2;
+
+    .fa-icon {
+      filter: drop-shadow(0 0 4px rgba(0 0 0 / 50%));
+      height: 1.75rem;
+      width: 1.75rem;
+    }
+  }
+
+  &-count {
+    align-items: center;
+    background-color: rgba(0 0 0 / 65%);
+    border-radius: 4px;
+    bottom: 0.5rem;
+    color: $text-color;
+    display: flex;
+    gap: 0.4rem;
+    padding: 0.15rem 0.6rem;
+    position: absolute;
+    right: 0.5rem;
+    z-index: 2;
+  }
+}
+
+button.Image {
+  background: none;
+  border: 0;
+  cursor: zoom-in;
+  padding: 0;
+
+  &:hover .Image-magnify {
+    opacity: 1;
+  }
+}
+
+// Top-level because Modal portals to body
+.ImageLightbox {
+  transition: transform 0.25s ease-out;
+
+  .modal.fade & {
+    transform: scale(0.97);
+  }
+
+  .modal.show & {
+    transform: none;
+  }
+
+  .modal-content {
+    backdrop-filter: blur(6px);
+    background-color: rgba(0 0 0 / 75%);
+  }
+
+  .modal-body {
+    display: flex;
+    gap: 1rem;
+    overflow: hidden;
+    padding: 1rem;
+  }
+
+  &-main {
+    align-items: center;
+    display: flex;
+    flex: 1;
+    flex-direction: column;
+    justify-content: center;
+    min-width: 0;
+    position: relative;
+
+    .Image {
+      flex: 1;
+      min-height: 0;
+    }
+
+    img {
+      object-fit: contain;
+      object-position: center;
+    }
+  }
+
+  &-caption {
+    color: $text-color;
+    padding-top: 0.5rem;
+  }
+
+  &-close {
+    position: absolute;
+    right: 0;
+    top: 0;
+    z-index: 2;
+
+    .fa-icon {
+      height: 2rem;
+      opacity: 0.6;
+      width: 2rem;
+    }
+
+    &:hover .fa-icon {
+      opacity: 1;
+    }
+  }
+
+  &-thumbs {
+    align-content: start;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+    overflow-y: auto;
+    padding-right: 0.75rem;
+    width: 500px;
+  }
+
+  &-thumb {
+    background: none;
+    border: 2px solid rgba(255 255 255 / 70%);
+    border-radius: 6px;
+    cursor: pointer;
+    flex-shrink: 0;
+    height: var(--thumb-height, 150px);
+    overflow: hidden;
+    padding: 0;
+    position: relative;
+    width: auto;
+
+    &.selected {
+      border-color: $danger;
+    }
+
+    img {
+      height: 100%;
+      object-fit: cover;
+      width: 100%;
+    }
+
+    &-dims {
+      background-color: rgba(0 0 0 / 65%);
+      border-radius: 3px 0 0;
+      bottom: 0;
+      color: $text-color;
+      font-size: 0.75rem;
+      padding: 0.1rem 0.4rem;
+      position: absolute;
+      right: 0;
+    }
+  }
 }
