--- frontend/src/components/imageGallery/index.ts
@@ -1,3 +1,0 @@
-import ImageGallery from "./ImageGallery";
-
-export default ImageGallery;
